+++ src/tilemap.cpp
@@ -130,8 +130,8 @@ tile* tilemap::getTileFromIndex( const sf::Vector2i &tileindex ) {
     return grid[tileindex.x][tileindex.y];
 }
 
-int tilemap::getTileSize(){
-    return int(tileSize);
+float tilemap::getTileSize(){
+    return tileSize;
 }
 
 sf::Vector2f tilemap::getMapPosition(){
